% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_caller.R
\name{run_caller}
\alias{run_caller}
\title{Run MiniCN copy-number caller}
\usage{
run_caller(
  sample_csv,
  outdir,
  genome_pkg = "BSgenome.Hsapiens.UCSC.hg38",
  panel_path = NULL,
  pc = 0.5,
  HIGH_AMP_T = log2(17/2),
  LOW_AMP_T = log2(7/2),
  GAIN_T = log2(3/2),
  LOSS_T = log2(1/2),
  DEEP_DEL_T = -2,
  MIN_AMPLICONS = 2L,
  MIN_GENE_Z = 3L,
  MAX_Q_VALUE = 0.05
)
}
\arguments{
\item{sample_csv}{CSV with columns: Tumor Folder, Tumor File Name, Normal Folder, Normal File Name}

\item{outdir}{output directory}

\item{genome_pkg}{BSgenome package name, (defualt "BSgenome.Hsapiens.UCSC.hg38")}

\item{panel_path}{path to panel gene list (HGNC symbols, required)}

\item{pc}{pseudocount for normalization (default 0.5)}

\item{HIGH_AMP_T}{high-amplification copy number threshold (default log2(17/2) or ~3.09)}

\item{LOW_AMP_T}{low-amplification copy number threshold (default log2(7/2) or ~1.81)}

\item{GAIN_T}{gain copy number threshold (default log2(3/2) or ~0.58)}

\item{LOSS_T}{loss copy number threshold (default log2(1/2) or -1)}

\item{DEEP_DEL_T}{deletion copy number threshold (default -2.0)}

\item{MIN_AMPLICONS}{minimum number of amplicons supporting respective changes (default 2)}

\item{MIN_GENE_Z}{minimum z-score cutoff for gains and losses (default 3)}

\item{MAX_Q_VALUE}{maximum q-value cutoff for gains and losses (default 0.05)}
}
\value{
merged tsv files and per sample pdf files
}
\description{
Run MiniCN copy-number caller
}
\examples{
# Package example files
sample_csv <- system.file("extdata", "sample_sheet.csv", package = "MiniCN")
panel_file <- system.file("extdata", "panel_genes.txt", package = "MiniCN")

# Output directory
out <- "outputs"
if (!dir.exists(out)) {
  dir.create(out, showWarnings = FALSE, recursive = TRUE)
}

# Skip if genome package is not available on this machine
if (requireNamespace("BSgenome.Hsapiens.UCSC.hg38", quietly = TRUE)) {
  run_caller(sample_csv = sample_csv, outdir = out, panel_path = panel_file)
}
}
